[Feat] 찜 정렬 완료

+++ UMC-Reborn/StoreHomeViewController/Models/JjimModel.swift
@@ -20,7 +20,7 @@ struct JjimListModel: Codable {
     let storeName: String
     let storeImage: String?
     let storeCategory: String
-    let storeScore: Int
+    let storeScore: Float
 }
